feat: lets package.json specify indexVersionLimit
