+++ src/util.ts
@@ -41,29 +41,3 @@ export namespace sort {
 }
 
 export const template = (context: any) => (t: string | undefined): string => _.template(t || '')(context)
-
-interface VersionsObject {
-  [key: string]: string;
-}
-
-export const sortVersionsObjectByKeysDesc = (input: VersionsObject): VersionsObject => {
-  const keys = Reflect.ownKeys(input).sort((a, b) => {
-    const splitA = (a as string).split('.').map(part => parseInt(part, 10))
-    const splitB = (b as string).split('.').map(part => parseInt(part, 10))
-    // sort by major
-    if (splitA[0] < splitB[0]) return 1
-    if (splitA[0] > splitB[0]) return -1
-    // sort by minor
-    if (splitA[1] < splitB[1]) return 1
-    if (splitA[1] > splitB[1]) return -1
-    // sort by patch
-    if (splitA[2] < splitB[2]) return 1
-    if (splitA[2] > splitB[2]) return -1
-    return 0
-  }) as string[]
-  const result: VersionsObject = {}
-  keys.forEach(key => {
-    result[key] = input[key]
-  })
-  return result
-}
